adapt linear_backend
