+++ python/sglang/srt/layers/attention/attention_registry.py
@@ -190,11 +190,13 @@ def attn_backend_wrapper(runner: "ModelRunner", full_attn_backend: "AttentionBac
         from sglang.srt.layers.attention.fla.utils import check_environments
         from sglang.srt.layers.attention.hybrid_linear_attn_backend import (
             HybridLinearAttnBackend,
-            LightningAttentionBackend,
             Mamba2AttnBackend,
         )
         from sglang.srt.layers.attention.linear.gdn_backend import GDNAttnBackend
         from sglang.srt.layers.attention.linear.kda_backend import KDAAttnBackend
+        from sglang.srt.layers.attention.linear.lightning_backend import (
+            LightningAttentionBackend,
+        )
         from sglang.srt.layers.attention.linear.utils import (
             initialize_linear_attn_config,
         )
